🚚(models) move ReferralActivity to a separate file
Move the ReferralActivity and its related enum model to a separate
file to clear up `referral.py`. As we were about to add new
values to the enum, we took this opportunity to split up the models
without affecting consumers as everything is re-exported from __init__.
