+++ src/backend/partaj/core/models/__init__.py
@@ -5,4 +5,5 @@ Make models.py a module containing one file per model as it was getting too long
 
 from .attachment import *
 from .referral import *
+from .referral_activity import *
 from .unit import *
